Implement the playGame function
this function is responsible for starting the game.
it initiates 5 round where the users choice is
compared against the computers choice. Then uses
keeps track of the winner of each round. at the end
of the five rounds, the player with the highest wins
the game. score variable are used to keep track
of each players score.